✨Move hardBan to Moderation, and hardban a user when permanent ban
